+++ CHANGES.rst
@@ -2,8 +2,16 @@ Changelog
 =========
 
 
+v0.0.3 (10-06-2017)
+*******************
+
+- Top Netlfix, HBO and Filmin
+- Recommendation from Netflix, HBO or Filmin
+- Fixed errors
+
+
 v0.0.2 (31-05-2017)
--------------------
+*******************
 
 - Search movies by title, year, director or cast.
 - Get the filmaffinity top and search by year
@@ -11,6 +19,6 @@ v0.0.2 (31-05-2017)
 
 
 v0.0.1 (29-05-2017)
--------------------
+*******************
 
 - Initial release.